+++ e2e/vue3-vite-e2e/tests/component.spec.ts
@@ -7,8 +7,6 @@ import {
   uniq,
 } from './utils';
 
-jest.setTimeout(60000);
-
 describe('component e2e', () => {
   let proj: string;